Slight change to narrative template
When resizing the narrative and moving from a small to a medium viewport,
the narrative text would not appear.  I've removed the conditional mobile
check on the template as the JS handles this anyway and to ensure that the
behaviour is as expected when viewed on a non mobile device.